chore: move the example to the integration directory
